refactor: Navbar and FooterSection to use MenuTitle component for menu rendering
